CalculateItemPrice working properly after test.

+++ src/main/java/org/js/checkoutcomponent/service/checkout/data/Item.java
@@ -2,10 +2,12 @@ package org.js.checkoutcomponent.service.checkout.data;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Item {
     private String id;
-    private double normalPrice;
-    private double specialPrice;
+    private BigDecimal normalPrice;
+    private BigDecimal specialPrice;
     private int requiredQuantity;
 }
